+++ src/components/Fencing/Fencing.tsx
@@ -6,7 +6,7 @@ import {UserPingRoomEvent} from '../../shared/roomEvents';
 import {useSocket} from '../../sockets/useSocket';
 import {emitAsync} from '../../sockets/emitAsync';
 import Player from '../Player';
-import {gameReducer, GameState, initialGameState} from '../../lib/reducers/gameV2';
+import {gameReducer, GameState, initialGameState} from '../../shared/gameEvents/reducers/gameReducer';
 import {GameEvent} from '../../shared/gameEvents';
 import {transformGameToPlayerProps} from './transformGameToPlayerProps';
 import {usePlayerActions} from './usePlayerActions';
