+++ src/components/Fencing/transformGameToPlayerProps.ts
@@ -1,7 +1,7 @@
 /**
  * Perhaps this whole file could live elsewhere, e.g. Player/transformGameToPlayerProps?
  * */
-import {GameJson} from '../../lib/reducers/gameV2';
+import {GameJson} from '../../shared/gameEvents/reducers/gameReducer';
 import {CluesJson} from '../../shared/types';
 import {CellCoords, CellIndex, Cursor, GridData, Ping} from '../Grid/types';
 import {PlayerActions} from './usePlayerActions';
