+++ src/shared/gameEvents/types/initializeGameEvent.ts
@@ -1,0 +1,4 @@
+import {GameEvent, GameEventType} from './GameEvent';
+
+export const isInitializeGameEvent = (event: GameEvent): event is GameEvent<GameEventType.initializeGame> =>
+  event.type === GameEventType.initializeGame;
